Rename folder to interfaces in remaining headers.
